+++ openquake/ghm/bin/create_curves.sh
@@ -3,7 +3,7 @@
 # This is the shell script used to run the code which produces an homogenised
 # set of hazard curves with global coverage. Note that it requires
 # the a-priori definition of two environment variables: REPOS and GEMDATA.
-# The former contains the `mosaic` repository, the latter is the path to the 
+# The former contains the `mosaic` repository, the latter is the path to the
 # repository containing the GEM dataset.
 #
 # Create the homogenised set of hazard curves
@@ -29,10 +29,12 @@ IMTSTR='PGA'
 # Shapefile with inland territories
 I_SHP='./../data/gis/inland.shp'
 #
-# List of models to be processed
-#MDLS="-m als,can,usa,mex,ucf,cca,sam"
+# List of models to be processed. This is an optional parameter. If not set,
+# i.e. MDLS="", all the models specified in `openquake.ghm.mosaic.DATA`
+# will be used.
+MDLS="-m als,can,usa,mex,ucf,cca,sam"
 #MDLS="-m mex,usa"
 #MDLS="-m cca,sam"
 #
 # Run hazard curves homogenisation
-../create_map_from_curves.py $C_SHP $O_PATH $D_PATH $SIDX $B_SHP $IMTSTR $I_SHP $MDLS
+../create_homogenised_curves.py $C_SHP $O_PATH $D_PATH $SIDX $B_SHP $IMTSTR $I_SHP $MDLS
